LayoutTests/webgpu/bind-group-layouts.html

<!DOCTYPE html><!-- webkit-test-runner [ experimental:WebGPUEnabled=true ] -->
<meta charset=utf-8>
<title>Create WebGPUBindGroupLayoutDescriptor.</title>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
function createBindGroupLayoutBinding() {
    return {
        binding: 0,
        visibility: WebGPUShaderStageBit.FRAGMENT | WebGPUShaderStageBit.VERTEX,
        type: "storageBuffer"
    };
}

test(() => {
    const bindGroupLayoutBinding = createBindGroupLayoutBinding();
    const bindGroupLayoutDescriptor = { bindings: [bindGroupLayoutBinding] };
    assert_not_equals(bindGroupLayoutDescriptor.bindings[0].visibility & WebGPUShaderStageBit.FRAGMENT, 0);
    assert_not_equals(bindGroupLayoutDescriptor.bindings[0].visibility & WebGPUShaderStageBit.VERTEX, 0);
    assert_equals(bindGroupLayoutDescriptor.bindings[0].type, "storageBuffer");
});

</script>
</body>
